renamed CCI Cryoclim parent dir
