+++ registry/data.txt
@@ -1,15 +1,17 @@
-WECANN/gpp_mon_WECANN-1-0_BE_gn_200701-201512.nc sha1:b98f8ac5d4786869d196eba1911e081200f2d6bc
-WECANN/hfls_mon_WECANN-1-0_BE_gn_200701-201512.nc sha1:62ed41d636dd6cbd8c238f54ca31e20a36ed09ab
-WECANN/hfss_mon_WECANN-1-0_BE_gn_200701-201512.nc sha1:9fe60acb368f50c9fb1f080eff8d3d6eea18b5ce
-LORA/mrro_mon_LORA_BE_gn_198001-201212.nc sha1:539dc49ac8db4a38f218163a0da0f8156f84fce4
-Hoffman/fgco2_yr_Hoffman_BE_gm_185007-201007.nc sha1:2456ff4394caf62afaed71ab065cf3f640e5bb5e
-Hoffman/nbp_yr_Hoffman_BE_gm_185007-201007.nc sha1:4d11fc76c8cc417dc78b432b5804a407b8340f4a
-WOA/no3_mon_WOA_BE_gn_201501-202212.nc sha1:4d0a18d12b32bbd0bdea94ea1ee369af32418f45
-WOA/o2_mon_WOA_BE_gn_201501-202212.nc sha1:b42689977b2202b682433d8f3c86cf5ff5d8f4ae
-WOA/po4_mon_WOA_BE_gn_201501-202212.nc sha1:fd60ba9cc79327c421bc4515f1b87c608f98ad26
-WOA/sio3_mon_WOA_BE_gn_201501-202212.nc sha1:9ed5bb211a321013f1b8bb03581b386abdf37410
-WOA/so_mon_WOA_BE_gn_201501-202212.nc sha1:56e2f86ee9000b563c60d9f1df8d614735ad9018
-WOA/thetao_mon_WOA_BE_gn_201501-202212.nc sha1:361416995c9886c0a98bd147c0c5a7d9d5e19740
-NOAA/ohc_yr_OHC_BE_gm_200506-202406.nc sha1:a918799d8e24e4f0015b9047a74d470ae9f0445c
-RAPID/amoc_mon_RAPID_BE_NA_200404-202302.nc sha1:e4ffb564f6448eebc8d0e251dcdc0f7ee107e179
-Snow-cci/snc_mon_Snow-cci_BE_gn_198201-201906.nc sha1:c0bfecd2f8b886e9301428d28bb6ff0507601be2
+CCI-CryoClim-FSC-1/snc_mon_Snow-cci_BE_gn_198201-201906.nc sha1:c0bfecd2f8b886e9301428d28bb6ff0507601be2
+WECANN-1-0/obs4MIPs_ColumbiaU_WECANN-1-0_mon_gpp_gn_v20250902.nc sha1:ada10d3d291008857461b97f2886dd18e7d02fe7
+WECANN-1-0/obs4MIPs_ColumbiaU_WECANN-1-0_mon_hfss_gn_v20250902.nc sha1:b5e7eb85ccf8a897b20781d30ee812ee13105e42
+WECANN-1-0/obs4MIPs_ColumbiaU_WECANN-1-0_mon_hfls_gn_v20250902.nc sha1:380cea38eb6adb8ff547376c06d83e296090d12d
+Hoffman-1-0/obs4MIPs_UCI-ORNL_Hoffman-1-0_yr_fgco2_gm_v20250903.nc sha1:f06d909348c05fdea0a7af87b01bfd4cb440e075
+Hoffman-1-0/obs4MIPs_UCI-ORNL_Hoffman-1-0_yr_nbp_gm_v20250903.nc sha1:803453d7b3873732190b1a20b25c4b50d34642c9
+HWSD-2-0/obs4MIPs_IIASA-FAO_HWSD-2-0_fx_cSoil_gn_v20250903.nc sha1:9855622a525902f0e2625ee6293e8d7c2b18c24c
+RAPID-2023-1a/obs4MIPs_NOC_RAPID-2023-1a_mon_msftmz_gm_v20250902.nc sha1:0dfeea11872bdc453cf5d0a3b0cf1abf4d14369f
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_monC_si_gn_v20250902.nc sha1:a51b67b0eb874ddd40ecb25cf9c087005186b74f
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_monC_thetao_gn_v20250902.nc sha1:ec62f9654f7effd2d0421121c709a49cf1728644
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_monC_no3_gn_v20250902.nc sha1:27d828bacc53e00fff3bd5af830592989a8de7ac
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_monC_o2_gn_v20250902.nc sha1:a0a9dc0a670777871bb5053a5c634e6f295f2401
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_monC_so_gn_v20250902.nc sha1:460d6a1d5b8164859b5f7a310dcd9823d0cb97ee
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_monC_po4_gn_v20250902.nc sha1:999257aa2be6aaaf03cae45c842e588c05fb7464
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_yr_ohc_gm_v20250903.nc sha1:dcf013e90b8b53813c614a01937a741a088cc0b9
+WOA-23/obs4MIPs_NOAA-NCEI-OCL_WOA-23_yr_ohcJm2_gn_v20250903.nc sha1:680980699b58a743abeab169ebb54556e3ba951f
+LORA-1-0/obs4MIPs_ARCCSS_LORA-1-0_mon_mrro_gn_v20250902.nc sha1:a62a9bb441a93114c578bfc8aab61fe4f38c1f71
